Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "plotly>=5.0.0",
]
